+++ src/main.jsx
@@ -5,7 +5,7 @@ import Root from './App.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    {/* <App /> */}
+    {/* { <App /> } */}
     <Root/>
   </React.StrictMode>,
 )
