+++ src/screens/show_screen/ShowScreen.css
